% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_id}
\alias{get_id}
\title{Extract a Single Subject's Data from a Grouped SleepCycle Object}
\usage{
get_id(sleepcycle_obj, .id)
}
\arguments{
\item{sleepcycle_obj}{An object of class \code{"SleepCycle"} that contains \strong{multiple IDs}.}

\item{.id}{The subject ID to extract from \code{sleepcycle_obj}. Must be present in the grouped dataset.}
}
\value{
A new \code{"SleepCycle"} object containing data \strong{only for the selected subject}, including:
\itemize{
\item \code{epoch}: Per-epoch sleep stage and cycle classifications.
\item \code{summary}: Sleep cycle summaries for the extracted subject.
\item \code{info}: Metadata about the \code{SleepCycle} object.
}
}
\description{
This function extracts the \strong{sleep cycle data} for a specific subject from a grouped \code{SleepCycle} object.
}
\details{
This function is used to \strong{extract a single subject's sleep cycle data} from a \strong{multi-subject dataset}.
If the provided \code{sleepcycle_obj} is \strong{not grouped}, the function stops execution to prevent unnecessary extraction.
}
