% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypnogram_single}
\alias{hypnogram_single}
\title{Example Hypnogram Datasets}
\format{
\itemize{
\item \code{hypnogram_single}: A \strong{data frame} with 954 rows and 2 columns:
\describe{
\item{epoch}{Sequential sleep epoch (numeric).}
\item{stage}{Sleep stage (character: "N3", "N2", "N1", "R", "W").}
}
\item \code{hypnogram_grouped}: A \strong{data frame} with 1912 rows and 3 columns:
\describe{
\item{epoch}{Sequential sleep epoch (numeric).}
\item{stage}{Sleep stage (character: "N3", "N2", "N1", "R", "W").}
\item{id}{Subject ID.}
}
}
}
\usage{
data(hypnogram_single)
data(hypnogram_grouped)
}
\description{
These datasets contain example hypnograms used for testing sleep cycle detection functions.
}
\details{
\itemize{
\item \code{hypnogram_single}: A \strong{single-subject} hypnogram.
\item \code{hypnogram_grouped}: A \strong{multi-subject} hypnogram with an \code{id} column.
}
}
\examples{
# Load the single-subject dataset
data(hypnogram_single)
head(hypnogram_single)

# Load the multi-subject dataset
data(hypnogram_grouped)
head(hypnogram_grouped)

# Use with sleepcycles_from_hypnogram()
sleepcycles_from_hypnogram(hypnogram_single, epoch_col = "epoch", stage_col = "stage")
sleepcycles_from_hypnogram(hypnogram_grouped, epoch_col = "epoch", stage_col = "stage", id_col = "id")
}
\keyword{datasets}
