% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepcycles_from_hypnogram.R
\name{print.SleepCycle}
\alias{print.SleepCycle}
\title{Custom Print Method for SleepCycle Objects}
\usage{
\method{print}{SleepCycle}(obj)
}
\arguments{
\item{obj}{A SleepCycle object.}
}
\description{
Custom Print Method for SleepCycle Objects
}
