% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ids}
\alias{plot_ids}
\alias{plot_cycle_proportions}
\alias{plot_cycle_counts}
\title{Grouped Sleep Cycle Visualization}
\usage{
plot_ids(sleepcycle_obj)

plot_cycle_proportions(sleepcycle_obj)

plot_cycle_counts(sleepcycle_obj)
}
\arguments{
\item{sleepcycle_obj}{An object of class \code{SleepCycle} that contains multiple IDs.}
}
\value{
A \code{ggplot} object.
}
\description{
These functions visualize sleep cycle statistics across multiple subjects.
}
\details{
\itemize{
\item \code{plot_ids()}: Displays sleep cycle progression across multiple IDs.
\item \code{plot_cycle_proportions()}: Shows the proportion of different sleep cycles.
\item \code{plot_cycle_counts()}: Plots a bar chart of cycle counts.
}
}
\examples{
plot_ids(sleepcycle_obj)
plot_cycle_proportions(sleepcycle_obj)
plot_cycle_counts(sleepcycle_obj)

}
