% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sleepstages}
\alias{sleepstages}
\title{Sleepstaging from one participant.}
\format{
A vmrk file (can be read as a text file) with 954 rows and 5 variables:
\describe{
\item{Type}{Describes what column 'Description' shows}
\item{Description}{Gives the sleep stage according to AASM criteria. 0 = wake, 1 = N1, 2 = N2, 3 = N3, 5 = W}
\item{Position}{Sampling point corresponding to the sleep stage marker at a sampling rate of 128 Hz.}
}
}
\usage{
sleepstages
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds. The variables are as follows:
}
\keyword{datasets}
