% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_summary}
\alias{plot_summary}
\title{Summary of Sleep Cycle Plots}
\usage{
plot_summary(sleepcycle_obj, id = NULL)
}
\arguments{
\item{sleepcycle_obj}{An object of class \code{SleepCycle}.}

\item{id}{(Optional) A subject identifier for grouped objects.}
}
\value{
A grid of \code{ggplot} objects.
}
\description{
This function generates a combined visualization of a hypnogram, sleep cycles, and, if using the 'dude' method for sleep cycles estimates, cycle densities.
}
\examples{
data("hypnogram_single")
plot_summary(
  sleepcycles_from_hypnogram(
    hypnogram_single,
    epoch_col = "epoch",
    stage_col = "stage",
    method = "dude",
    verbose = FALSE
  )
)
plot_summary(
  sleepcycles_from_hypnogram(
    hypnogram_single,
    epoch_col = "epoch",
    stage_col = "stage",
    method = "feinberg",
    verbose = FALSE
  )
)
}
