% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepcycles_from_hypnogram.R
\name{print.SleepCycle}
\alias{print.SleepCycle}
\title{Custom Print Method for SleepCycle Objects}
\usage{
\method{print}{SleepCycle}(x, ...)
}
\arguments{
\item{x}{A SleepCycle object.}

\item{...}{not currently implemented.}
}
\description{
Custom Print Method for SleepCycle Objects
}
