% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hypnogram}
\alias{plot_hypnogram}
\alias{plot_densities}
\alias{plot_cycles}
\title{Sleep Cycle Visualization Functions}
\usage{
plot_hypnogram(
  sleepcycle_obj,
  id = NULL,
  stage_order = NULL,
  overlay_cycles = TRUE,
  overlay_clrs = c("#3B528BFF", "#5DC863FF")
)

plot_densities(
  sleepcycle_obj,
  id = NULL,
  include_levels = NULL,
  clrs = c("#3B528BFF", "#5DC863FF"),
  overlay_cycles = TRUE,
  overlay_clrs = clrs
)

plot_cycles(sleepcycle_obj, id = NULL)
}
\arguments{
\item{sleepcycle_obj}{An object of class \code{SleepCycle}.}

\item{id}{(Optional) A subject identifier for grouped objects.}

\item{stage_order}{(Optional) Reorders sleep stages in \code{plot_hypnogram()}.}

\item{overlay_cycles}{Logical. If \code{TRUE}, overlays sleep cycles on the plots.}

\item{overlay_clrs}{Color palette for cycle overlays.}

\item{include_levels}{(Optional) Specifies which density levels to plot in \code{plot_densities()}.}

\item{clrs}{Color palette for different sleep states.}
}
\value{
A \code{ggplot} object.
}
\description{
These functions generate plots related to sleep cycles from a \code{SleepCycle} object.
}
\details{
\itemize{
\item \code{plot_hypnogram()}: Creates a hypnogram of sleep stages over time.
\item \code{plot_densities()}: Plots density estimates of sleep stages.
\item \code{plot_cycles()}: Visualizes the structure of sleep cycles.
}
}
